+++ RELEASENOTES.md
@@ -43,6 +43,12 @@
         ([#99](https://github.com/androidx/media/issues/99)).
     *   `SimpleBitmapLoader` can load bitmap from `file://` URIs
         ([#108](https://github.com/androidx/media/issues/108)).
+    *   Fix bug where the `MediaSessionService` was kept started in the
+        foreground and a notification with a pause button was still shown when
+        the player completed playback. Now, once playback is completed, the
+        service is stopped from the foreground and a notification with a play
+        button is shown to restart playback of the last media item
+        ([#112](https://github.com/androidx/media/issues/112)).
 *   RTSP:
     *   Add H263 fragmented packet handling
         ([#119](https://github.com/androidx/media/pull/119)).